+++ src/mempool.c
@@ -653,9 +653,9 @@ mp_pool_garbage_collect(void *arg)
 void
 mp_pool_log_status(mp_pool_t *pool)
 {
-  uint64_t bytes_used = 0;
-  uint64_t bytes_allocated = 0;
-  uint64_t bu = 0, ba = 0;
+  unsigned long long bytes_used = 0;
+  unsigned long long bytes_allocated = 0;
+  unsigned long long bu = 0, ba = 0;
   mp_chunk_t *chunk;
   int n_full = 0, n_used = 0;
 
@@ -664,7 +664,7 @@ mp_pool_log_status(mp_pool_t *pool)
   for (chunk = pool->empty_chunks; chunk; chunk = chunk->next)
     bytes_allocated += chunk->mem_size;
 
-  ircd_log(LOG_DBG, "%lu bytes in %d empty chunks",
+  ircd_log(LOG_DBG, "%llu bytes in %d empty chunks",
        bytes_allocated, pool->n_empty_chunks);
   for (chunk = pool->used_chunks; chunk; chunk = chunk->next) {
     ++n_used;
@@ -675,7 +675,7 @@ mp_pool_log_status(mp_pool_t *pool)
          chunk->n_allocated);
   }
 
-  ircd_log(LOG_DBG, "%lu/%lu bytes in %d partially full chunks",
+  ircd_log(LOG_DBG, "%llu/%llu bytes in %d partially full chunks",
        bu, ba, n_used);
   bytes_used += bu;
   bytes_allocated += ba;
@@ -687,21 +687,21 @@ mp_pool_log_status(mp_pool_t *pool)
     ba += chunk->mem_size;
   }
 
-  ircd_log(LOG_DBG, "%lu/%lu bytes in %d full chunks",
+  ircd_log(LOG_DBG, "%llu/%llu bytes in %d full chunks",
        bu, ba, n_full);
   bytes_used += bu;
   bytes_allocated += ba;
 
-  ircd_log(LOG_DBG, "Total: %lu/%lu bytes allocated "
+  ircd_log(LOG_DBG, "Total: %llu/%llu bytes allocated "
        "for cell pools are full.",
        bytes_used, bytes_allocated);
 
 #ifdef MEMPOOL_STATS
-  ircd_log(LOG_DBG, "%lu cell allocations ever; "
-       "%lu chunk allocations ever; "
-       "%lu chunk frees ever.",
-       pool->total_items_allocated,
-       pool->total_chunks_allocated,
-       pool->total_chunks_freed);
+  ircd_log(LOG_DBG, "%llu cell allocations ever; "
+       "%llu chunk allocations ever; "
+       "%llu chunk frees ever.",
+       (long long)pool->total_items_allocated,
+       (long long)pool->total_chunks_allocated,
+       (long long)pool->total_chunks_freed);
 #endif
 }
